test: split audit-row tests into their own file (max-lines gate)
