+++ server/tests/integration/test_cloud_integration_gateway_api.py
@@ -468,169 +468,3 @@ async def test_list_tools_cache_refetches_after_ttl(
     third = await client.post(GATEWAY_URL, headers=headers, json=payload)
     assert third.status_code == 200, third.text
     assert calls["count"] == 2
-
-
-async def _tool_call_events(db_session: AsyncSession):
-    from sqlalchemy import select
-
-    from proliferate.db.models.cloud.integrations import CloudIntegrationToolCallEvent
-
-    return (
-        (
-            await db_session.execute(
-                select(CloudIntegrationToolCallEvent).order_by(
-                    CloudIntegrationToolCallEvent.created_at
-                )
-            )
-        )
-        .scalars()
-        .all()
-    )
-
-
-@pytest.mark.asyncio
-async def test_call_tool_success_writes_audit_event(
-    client: AsyncClient, db_session: AsyncSession, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    bearer = await _gateway_bearer(client, db_session, prefix="gw-audit-ok")
-    headers = {"Authorization": f"Bearer {bearer}"}
-
-    from sqlalchemy import select
-
-    from proliferate.db.models.cloud.runtime_workers import CloudRuntimeWorker
-
-    worker = (await db_session.execute(select(CloudRuntimeWorker))).scalars().first()
-    assert worker is not None
-    await _seed_ready_account(db_session, user_id=str(worker.owner_user_id), namespace="context7")
-
-    async def fake_call_tool(*, url, headers, tool_name, arguments, query=None):
-        return {"content": [{"type": "text", "text": "ok"}], "isError": False}
-
-    monkeypatch.setattr(
-        "proliferate.server.cloud.integration_gateway.service.mcp_remote.call_tool",
-        fake_call_tool,
-    )
-
-    result = await _tool_call(
-        client,
-        headers,
-        name="integrations.call_tool",
-        arguments={"provider": "context7", "tool": "resolve-library-id", "arguments": {}},
-    )
-    assert result["structuredContent"]["isError"] is False
-
-    events = await _tool_call_events(db_session)
-    assert len(events) == 1
-    event = events[0]
-    assert event.ok is True
-    assert event.error_code is None
-    assert event.integration_namespace == "context7"
-    assert event.tool_name == "resolve-library-id"
-    assert event.user_id == worker.owner_user_id
-    assert event.organization_id is None
-    assert event.runtime_worker_id == worker.id
-    assert event.latency_ms >= 0
-
-
-@pytest.mark.asyncio
-async def test_call_tool_upstream_failure_writes_audit_event(
-    client: AsyncClient, db_session: AsyncSession, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    bearer = await _gateway_bearer(client, db_session, prefix="gw-audit-down")
-    headers = {"Authorization": f"Bearer {bearer}"}
-
-    from sqlalchemy import select
-
-    from proliferate.db.models.cloud.runtime_workers import CloudRuntimeWorker
-    from proliferate.integrations.mcp_remote import McpRemoteError
-
-    worker = (await db_session.execute(select(CloudRuntimeWorker))).scalars().first()
-    assert worker is not None
-    await _seed_ready_account(db_session, user_id=str(worker.owner_user_id), namespace="context7")
-
-    async def failing_call_tool(*, url, headers, tool_name, arguments, query=None):
-        raise McpRemoteError("upstream is down", code="transport_error")
-
-    monkeypatch.setattr(
-        "proliferate.server.cloud.integration_gateway.service.mcp_remote.call_tool",
-        failing_call_tool,
-    )
-
-    result = await _tool_call(
-        client,
-        headers,
-        name="integrations.call_tool",
-        arguments={"provider": "context7", "tool": "resolve-library-id", "arguments": {}},
-    )
-    assert result["isError"] is True
-
-    events = await _tool_call_events(db_session)
-    assert len(events) == 1
-    event = events[0]
-    assert event.ok is False
-    assert event.error_code == "transport_error"
-    assert event.integration_namespace == "context7"
-    assert event.tool_name == "resolve-library-id"
-
-
-@pytest.mark.asyncio
-async def test_call_tool_tool_level_error_writes_audit_event(
-    client: AsyncClient, db_session: AsyncSession, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    # The upstream answered, but the tool itself reported an error: still
-    # audited, as a failure with the dedicated tool_error code.
-    bearer = await _gateway_bearer(client, db_session, prefix="gw-audit-toolerr")
-    headers = {"Authorization": f"Bearer {bearer}"}
-
-    from sqlalchemy import select
-
-    from proliferate.db.models.cloud.runtime_workers import CloudRuntimeWorker
-
-    worker = (await db_session.execute(select(CloudRuntimeWorker))).scalars().first()
-    assert worker is not None
-    await _seed_ready_account(db_session, user_id=str(worker.owner_user_id), namespace="context7")
-
-    async def erroring_call_tool(*, url, headers, tool_name, arguments, query=None):
-        return {"content": [{"type": "text", "text": "bad args"}], "isError": True}
-
-    monkeypatch.setattr(
-        "proliferate.server.cloud.integration_gateway.service.mcp_remote.call_tool",
-        erroring_call_tool,
-    )
-
-    result = await _tool_call(
-        client,
-        headers,
-        name="integrations.call_tool",
-        arguments={"provider": "context7", "tool": "resolve-library-id", "arguments": {}},
-    )
-    assert result["structuredContent"]["isError"] is True
-
-    events = await _tool_call_events(db_session)
-    assert len(events) == 1
-    assert events[0].ok is False
-    assert events[0].error_code == "tool_error"
-
-
-@pytest.mark.asyncio
-async def test_call_tool_unknown_provider_writes_audit_event(
-    client: AsyncClient, db_session: AsyncSession
-) -> None:
-    bearer = await _gateway_bearer(client, db_session, prefix="gw-audit-noprov")
-    headers = {"Authorization": f"Bearer {bearer}"}
-
-    result = await _tool_call(
-        client,
-        headers,
-        name="integrations.call_tool",
-        arguments={"provider": "nope", "tool": "x", "arguments": {}},
-    )
-    assert result["isError"] is True
-
-    events = await _tool_call_events(db_session)
-    assert len(events) == 1
-    event = events[0]
-    assert event.ok is False
-    assert event.error_code == "integration_provider_not_found"
-    assert event.integration_namespace == "nope"
-    assert event.tool_name == "x"
